refactor(aaa): modularise migration
